requirements.txt
locust==2.20.0
grpcio==1.60.0
grpcio-tools==1.60.0